Move models out to src/python
