+++ src/python/api/gql/types/user_activities.py
@@ -14,7 +14,7 @@ from sqlalchemy.sql.functions import now
 from typing import Any, Type
 
 from ....config import db
-from ....api.models import UserActivity
+from ....models import UserActivity
 
 
 def user_activity_fields() -> dict[str, GraphQLField]:
