+++ src/python/models/subscription_notification.py
@@ -1,4 +1,4 @@
-from ...config import db
+from ..config import db
 import datetime
